continue to remove concept of current time
