+++ src/App.tsx
@@ -20,11 +20,20 @@ import { Feature, Geometry, GeoJsonProperties } from 'geojson';
 import Control from 'react-leaflet-custom-control';
 import toDTG from './helpers/toDTG.ts';
 import { AppDispatch } from './app/store.ts';
-import { TimeSteps } from './helpers/generateCurrentLocations.ts';
 
 interface FileHandler {
   blobType: string
   handle: (text: string, features: Feature<Geometry, GeoJsonProperties>[], dispatch: AppDispatch) => void
+}
+
+export interface TimeState {
+  start: number;
+  step: string;
+  end: number;
+}
+
+export const timeVal = (timeStr: string): number => {
+  return new Date(timeStr).getTime()
 }
 
 const FileHandlers: FileHandler[] = [
@@ -69,7 +78,7 @@ function App() {
       timeInitialised.current = true
       const timeBounds = timeBoundsFor(features)
       setTimeBounds(timeBounds)
-      const timePayload = { start: timeBounds[0], step: TimeSteps[1], current: (timeBounds[0] + timeBounds[1]) / 2, end: timeBounds[1] }
+      const timePayload = { start: timeBounds[0], step: '00h30m', end: timeBounds[1] }
       setTime(timePayload)
     }
   }, [features])
