--- src/helpers/generateCurrentLocations.ts
@@ -1,78 +1,0 @@
-import * as turf from "@turf/turf";
-import { Feature, MultiPoint, Point, Position } from "geojson";
-
-export interface TimeState {
-  start: number;
-  current: number;
-  step: string;
-  end: number;
-}
-
-export const TimeSteps = ['00h15m','00h30m', '01h00m', '02h00m', '03h00m', '06h00m']
-
-const isTemporal = (feature: Feature): boolean => {
-  return feature.properties?.times
-}
-
-export const timeVal = (timeStr: string): number => {
-  return new Date(timeStr).getTime()
-}
-
-export const calcInterpLocation = (poly: MultiPoint, times: string[], current: number): Position | undefined => {
-  const index = times.findIndex((time: string) => timeVal(time) >= current)
-  if (index == 0) {
-    // check if it's the same time as the first point
-    const firstTime = timeVal(times[0])
-    return (firstTime === current) ? poly.coordinates[0] : undefined
-  } else if (index > 0) {
-    const coords = poly.coordinates
-    const isFirst = index === 0
-    const beforeIndex = isFirst ? 0 : index - 1
-    const afterIndex = isFirst ? 0 : index
-    const beforeCoords = coords[beforeIndex]
-    const afterCoords = coords[afterIndex]
-    const before = turf.point(beforeCoords)
-    const after = turf.point(afterCoords)
-    const turfPath = turf.lineString([beforeCoords, afterCoords])
-    const len = turf.distance(before, after)
-    const beforeTime = timeVal(times[beforeIndex])
-    const afterTime = timeVal(times[afterIndex])
-    const timeDelta = afterTime - beforeTime
-    const proportion = (current - beforeTime) / timeDelta
-    const lenProp = len * proportion
-    const interpolated = isNaN(lenProp) ? before : turf.along(turfPath, lenProp)
-    const markerLoc = interpolated.geometry.coordinates
-    return markerLoc
-  } else {
-    return undefined
-  }
-}
-
-export const generateCurrentLocations = (features: Feature[], time: TimeState): Feature<Point>[] => {
-  const temporalFeatures = features.filter(isTemporal)
-  const pointFeatures = temporalFeatures.map((feature) => {
-    const times = feature.properties?.times
-    const poly = feature.geometry as MultiPoint
-    const markerLoc = calcInterpLocation(poly, times, time.current)
-    if (markerLoc) {
-      const pointFeature: Feature<Point> = {
-        type: 'Feature',
-        id: feature.id,
-        geometry: {
-          type: 'Point',
-          coordinates: markerLoc
-        },
-        properties: {
-          name: feature.properties?.name,
-          color: feature.properties?.color
-        }
-      }
-      return pointFeature
-    } else {
-      return undefined
-    }
-  })
-  const res = pointFeatures.filter((f) => f !== undefined) as Feature<Point>[]
-  return res
-}
-
